+++ src/core/sagas/Sagas.js
@@ -42,7 +42,6 @@ export default function* sagas() :Generator<*, *, *> {
     fork(StudiesSagas.getParticipantsEnrollmentStatusWatcher),
     fork(StudiesSagas.getStudiesWatcher),
     fork(StudiesSagas.getStudyParticipantsWatcher),
-    fork(StudiesSagas.getStudyAuthorizationsWatcher),
     fork(StudiesSagas.updateStudyWatcher),
 
     // DataSagas
